requirements.txt
apscheduler
flask
flask-socketio
paho-mqtt
scipy>=0.17
pyyaml
jinja2
ruamel.yaml